requirements.txt
tensorflow
numpy
streamlit
Pillow
scikit-learn
matplotlib